REF: move BeamsService from bin to beams.service.rpc_handler, where it is relevant
